Implementation of a wave system
Inside the inspector we can setup a list of enemies to spawn and their positions.

A coroutine will spawn every waves, and between each waves will wait if enemies are still alive or not.

We need to pre intanciate the enemies because reactivating the enemies in the pool is faster than instanciating them, this mean they will have a difference is their movement.